+++ mpv.net/Misc/Misc.cs
@@ -3,7 +3,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Drawing;
 using System.IO;
 using System.Linq;
@@ -16,100 +15,6 @@ using Microsoft.Win32;
 
 namespace mpvnet
 {
-    public class App
-    {
-        public static string RegPath { get; } = @"HKCU\Software\" + Application.ProductName;
-        public static string ConfPath { get; } = mp.ConfigFolder + "\\mpvnet.conf";
-        public static string DarkMode { get; set; } = "always";
-        public static string ProcessInstance { get; set; } = "single";
-        public static string DarkColor { get; set; }
-        public static string LightColor { get; set; }
-
-        public static string[] VideoTypes { get; } = "264 265 asf avc avi avs flv h264 h265 hevc m2ts m2v m4v mkv mov mp4 mpeg mpg mpv mts ts vob vpy webm webm wmv y4m".Split(' ');
-        public static string[] AudioTypes { get; } = "mp3 mp2 ac3 ogg opus flac wav w64 m4a dts dtsma dtshr dtshd eac3 thd thd+ac3 mka aac mpa".Split(' ');
-        public static string[] ImageTypes { get; } = "jpg bmp gif png".Split(' ');
-        public static string[] SubtitleTypes { get; } = "srt ass idx sup ttxt ssa smi".Split(' ');
-        public static string[] UrlWhitelist { get; set; } = { "tube", "vimeo", "ard", "zdf" };
-
-        public static bool RememberHeight { get; set; } = true;
-        public static bool RememberPosition { get; set; }
-        public static bool DebugMode { get; set; }
-        public static bool IsStartedFromTerminal { get; } = Environment.GetEnvironmentVariable("_started_from_console") == "yes";
-
-        public static int StartThreshold { get; set; } = 1500;
-
-        public static bool IsDarkMode {
-            get => (DarkMode == "system" && Sys.IsDarkTheme) || DarkMode == "always";
-        }
-
-        public static void Init()
-        {
-            string dummy = mp.ConfigFolder;
-            var dummy2 = mp.Conf;
-
-            foreach (var i in Conf)
-                ProcessProperty(i.Key, i.Value);
-
-            if (App.DebugMode)
-            {
-                try
-                {
-                    string filePath = mp.ConfigFolder + "\\mpvnet-debug.log";
-                    if (File.Exists(filePath)) File.Delete(filePath);
-                    Trace.Listeners.Add(new TextWriterTraceListener(filePath));
-                    Trace.AutoFlush = true;
-                    //if (App.DebugMode) Trace.WriteLine("");
-                }
-                catch (Exception e)
-                {
-                    Msg.ShowException(e);
-                }
-            }
-        }
-
-        static Dictionary<string, string> _Conf;
-
-        public static Dictionary<string, string> Conf {
-            get {
-                if (_Conf == null)
-                {
-                    _Conf = new Dictionary<string, string>();
-
-                    if (File.Exists(ConfPath))
-                        foreach (string i in File.ReadAllLines(ConfPath))
-                            if (i.Contains("=") && !i.StartsWith("#"))
-                                _Conf[i.Substring(0, i.IndexOf("=")).Trim()] = i.Substring(i.IndexOf("=") + 1).Trim();
-                }
-                return _Conf;
-            }
-        }
-
-        public static void UnknownModule(string path)
-        {
-            Msg.ShowError("Failed to load script or extension", "Only scripts and extensions that ship with mpv.net are allowed in <startup>\\scripts or <startup>\\extensions.\n\nUser scripts or extensions have to use <config folder>\\scripts or <config folder>\\extensions.\n\nNever copy a new mpv.net version over a old mpv.net version.\n\nNever install a new mpv.net version on top of a old mpv.net version.\n\n" + path);
-        }
-
-        public static bool ProcessProperty(string name, string value)
-        {
-            switch (name) // return true instead of break!
-            {
-                case "remember-position": RememberPosition = value == "yes"; return true;
-                case "start-size": RememberHeight = value == "previous"; return true;
-                case "process-instance": ProcessInstance = value; return true;
-                case "dark-mode": DarkMode = value; return true;
-                case "debug-mode": DebugMode = value == "yes"; return true;
-                case "dark-color": DarkColor = value.Trim('\'', '"'); return true;
-                case "light-color": LightColor = value.Trim('\'', '"'); return true;
-                case "url-whitelist": UrlWhitelist = value.Split(' ', ',', ';'); return true;
-                case "start-threshold":
-                    int.TryParse(value, out int result);
-                    StartThreshold = result;
-                    return true;
-            }
-            return false;
-        }
-    }
-
     public class Sys
     {
         public static bool IsDarkTheme {